+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - JVM excludes via vm arguments (like `steadybit.agent.disable-jvm-attachment`) are working again
 - Option to validate user provided class and method name for "Java Method Delay" and" "Java Method Exception" attacks
+- Align method parameter of "Controller Exception" and "Controller Delay" to "HTTP Client Status" and accept multiple values
 
 ## v1.1.8